first draft of GPS (Ublox Neo) support in oXs
In order to activate the GPS option, the file oxs_config.h must contain
a line with : #define GPS_INSTALLED.
Tx pin of GPS must be connected to Rx pin of Arduino
Rx pin of GPS must ne connected to pin 6 of arduino (best is to add a
resistor e.g. 10 k in serie in order to protect GPS because Arduino pin
is 6 volt).
There is no need to specify which GPS data have to be transmitted. When
a GPS is installed, 5 data are always transmitted (longitude, latitude,
altitude, 3d speed and course). Note data are transmitted only when GPS
has a 3D fix. oXs uses therefore the device ID foreseen by FRSKY for a
GPS device.