HoTT bugfix
After receiving a request of the form 0x80 - not_our_own_id, the old
code delayed further reception for 4ms.  This causes a problem when the
receiver is turned on and the transmitter is not: The receiver will
issue requests 0x80 - 0x80 in that case. The very first request of this
form is handled as expected.  When the first byte of the second request
is received, the code still remembers the last received byte (0x80) and
thus immediately stops reception for 4ms, thereby missing the second
byte of that request.  This repeats for all further requests.  Even when
the transmitter is eventually turned on and the receiver starts issuing
requests for other sensor IDs, the code will always ignore the second
request byte and will thus not be able to transmit any telemetry.

The problem can be solved by either forcing it to forget the previous
byte in case some other than our own sensor ID is received, or by
omitting the 4ms delay until accepting further data from the receiver.
Here the latter option was chosen.